requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
mysql-connector-python==8.1.0
pydantic==2.4.2
python-dotenv==1.0.0
jinja2==3.1.2